refactor: split CLI chat command
